Move platform keycode code into platform modules
