+++ druid-shell/src/platform/mac/mod.rs
@@ -18,6 +18,7 @@
 
 pub mod application;
 pub mod dialog;
+pub mod keycodes;
 pub mod menu;
 pub mod runloop;
 pub mod util;
