+++ druid-shell/src/platform/mac/window.rs
@@ -45,7 +45,8 @@ use super::util::{assert_main_thread, make_nsstring};
 use crate::clipboard::ClipboardItem;
 use crate::common_util::IdleCallback;
 use crate::dialog::FileDialogOptions;
-use crate::keyboard::{KeyCode, KeyEvent, KeyModifiers};
+use crate::keyboard::{KeyEvent, KeyModifiers};
+use crate::keycodes::KeyCode;
 use crate::mouse::{Cursor, MouseButton, MouseEvent};
 use crate::platform::application::Application;
 use crate::window::{FileInfo, Text, TimerToken, WinCtx, WinHandler};
